feat(node)!: replace async module.register with sync module.registerHooks
Merge sync.mts implementation into hooks.mts and update register.mts
to use module.registerHooks() directly. Remove the separate ./sync
export as the main entry point is now sync-only.

BREAKING CHANGE: async hooks via module.register() are replaced with
sync hooks via module.registerHooks(). Requires Node.js >= 22.18.0.
