+++ packages/node/package.json
@@ -16,7 +16,7 @@
   },
   "license": "MIT",
   "engines": {
-    "node": ">=22.14.0"
+    "node": ">=22.18.0"
   },
   "type": "module",
   "main": "dist/register.mjs",
@@ -31,18 +31,6 @@
         "types": "./dist/register.d.mts",
         "dev": "./src/register.mts",
         "default": "./dist/register.mjs"
-      }
-    },
-    "./sync": {
-      "module-sync": {
-        "types": "./dist/sync.d.mts",
-        "dev": "./src/sync.mts",
-        "default": "./dist/sync.mjs"
-      },
-      "import": {
-        "types": "./dist/sync.d.mts",
-        "dev": "./src/sync.mts",
-        "default": "./dist/sync.mjs"
       }
     },
     "./hooks": {
@@ -87,8 +75,6 @@
   "scripts": {
     "build:clean": "rimraf dist && rm -f tsconfig.tsbuildinfo",
     "example3": "node --experimental-config-file=node.config.json --test examples/bowling.test.mts",
-    "example:sync2": "node --conditions=dev --enable-source-maps --import @power-assert/node/sync --test examples/bowling.test.mts",
-    "example:sync": "node --conditions=dev --import @power-assert/node/sync --test examples/bowling.test.mts",
     "example:ts": "node --conditions=dev --experimental-strip-types --import @power-assert/node --test examples/bowling.test.mts",
     "example": "node --conditions=dev --enable-source-maps --import @power-assert/node --test examples/bowling.test.mjs",
     "demo": "node --conditions=dev --enable-source-maps --import @power-assert/node --test examples/demo.test.mjs",
